% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PositionalIsomeRs.r
\name{PositionalIsomeRs}
\alias{PositionalIsomeRs}
\title{PositionalIsomeRs}
\usage{
PositionalIsomeRs(
  c1 = 107.9,
  c2 = 108.5,
  c3 = 108.9,
  c4 = 109.3,
  f1 = "150.078[0-9]",
  f2 = "282.121[0-9]",
  intfilt = 50000
)
}
\arguments{
\item{c1}{first cutoff - 108 lower limit}

\item{c2}{second cutoff - 108 upper limit}

\item{c3}{third cutoff - 109 lower limit}

\item{c4}{fourth cutoff -109 upper limit}

\item{f1}{filter value}

\item{f2}{filter value}

\item{intfilt}{intensity filter upper limit}
}
\value{
None
}
\description{
Processes multiple ms3 files in the current directory.
Scans in the specified MS3 files are filtered by the ms1 full scan parameters.
From each of the matching scans the masses closest to two user specified values
(default value 108 and 109) are identified to compare the intensity values.
Plots are created to visualize the intensity vs the retention time for both specified masses.
Plots are also created to compare the log2 ratio of the intensity vs the mass.
}
\examples{
PositionalIsomeRs()
PositionalIsomeRs(107.9,108.5,108.9,109.3, "150.078[0-9]", "282.121[0-9]")

}
