% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMS3.r
\name{readMS3}
\alias{readMS3}
\title{readMS3}
\usage{
readMS3(file.in)
}
\arguments{
\item{file.in}{file name of .ms3 file}
}
\value{
a list with scan no, retention time, filter & MassInt
}
\description{
Reads an .ms3 file
}
\examples{
readMS3("10mix_1.ms3")

}
