% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallMolecule.r
\name{smallMolecule}
\alias{smallMolecule}
\title{smallMolecule}
\usage{
smallMolecule(data.out, cutoffs, filter, int.filter = NA, debug.test = F)
}
\arguments{
\item{cutoffs}{mass cutoff values}

\item{filter}{filter values}

\item{int.filter}{intensity upper limit}

\item{debug.test}{set to TRUE for debugging output}

\item{data.in}{list structure produced by readMS3}
}
\value{
a data frame of filtered values
}
\description{
Scans in the specified MS3 files are filtered by the ms1 full scan parameters.
From each of the matching scans the masses closest to two user specified values
(default value 108 and 109) are identified to compare the intensity values.
Plots are created to visualize the intensity vs the retention time for both specified masses.
Plots are also created to compare the log2 ratio of the intensity vs the mass.
}
\examples{
smallMolecule(data.out, cutoffs=c(c1,c2,c3,c4), filter=c(f1,f2))

}
